move decode_byte_string out of tests
