+++ solana/utils/helpers.py
@@ -1,5 +1,8 @@
 """Helper functions."""
 
+from base64 import b64decode
+from base58 import b58decode
+
 
 def from_uint8_bytes(uint8: bytes) -> int:
     """Convert from uint8 to python int."""
@@ -9,3 +12,14 @@ def from_uint8_bytes(uint8: bytes) -> int:
 def to_uint8_bytes(val: int) -> bytes:
     """Convert an integer to uint8."""
     return val.to_bytes(1, byteorder="little")
+
+
+def decode_byte_string(byte_string: str, encoding: str = "base64") -> bytes:
+    """Decode a encoded string from an RPC Response."""
+    b_str = str.encode(byte_string)
+    if encoding == "base64":
+        return b64decode(b_str)
+    if encoding == "base58":
+        return b58decode(b_str)
+
+    raise NotImplementedError(f"{encoding} decoding not currently supported.")
