+++ tests/integration/test_token_client.py
@@ -5,10 +5,11 @@ import pytest
 import spl.token._layouts as layouts
 from solana.publickey import PublicKey
 from solana.rpc.types import TxOpts
+from solana.utils.helpers import decode_byte_string
 from spl.token.client import Token
 from spl.token.constants import ASSOCIATED_TOKEN_PROGRAM_ID, TOKEN_PROGRAM_ID
 
-from .utils import AIRDROP_AMOUNT, assert_valid_response, confirm_transaction, decode_byte_string
+from .utils import AIRDROP_AMOUNT, assert_valid_response, confirm_transaction
 
 
 @pytest.mark.integration
