+++ tests/integration/utils.py
@@ -1,10 +1,7 @@
 """Integration test utils."""
 import asyncio
 import time
-from base64 import b64decode
 from typing import Any, Dict
-
-from base58 import b58decode
 
 from solana.rpc.api import Client
 from solana.rpc.async_api import AsyncClient
@@ -99,14 +96,3 @@ async def aconfirm_transaction(client: AsyncClient, tx_sig: str) -> RPCResponse:
     if not resp["result"]:
         raise RuntimeError("could not confirm transaction: ", tx_sig)
     return resp
-
-
-def decode_byte_string(byte_string: str, encoding: str = "base64") -> bytes:
-    """Decode a encoded string from an RPC Response."""
-    b_str = str.encode(byte_string)
-    if encoding == "base64":
-        return b64decode(b_str)
-    if encoding == "base58":
-        return b58decode(b_str)
-
-    raise NotImplementedError(f"{encoding} decoding not currently supported.")
